Remove workspace scoping in route
